chores(refactor): refactor - Refactor Landing Page Structure: Split page components as section - Refactor Appbar component, split some components as UI component - Delete Unused files
