+++ client/src/axios-instance.js
@@ -1,8 +1,8 @@
 import axios from "axios";
 
-const cyclicServer = 'https://clumsy-pink-bedclothes.cyclic.app/api/'
-// const localServer = 'http://localhost:5000/api/'
+// const cyclicServer = 'https://clumsy-pink-bedclothes.cyclic.app/api/'
+const localServer = 'http://localhost:5000/api/'
 const instance = axios.create({
-    baseURL: cyclicServer,
+    baseURL: localServer,
   });
   export default instance;
