+++ client/src/pages/login-page.js
@@ -1,7 +1,6 @@
 import LoginForm from "../components/UI/Forms/login-form";
 import React from "react";
 import { Box } from "@mui/material";
-import Header from "../components/UI/Navbar/Header";
 import CustomAppbar from "../components/UI/Appbar/custom-appbar";
 
 export const LoginPage = () => {
@@ -11,7 +10,7 @@ export const LoginPage = () => {
         height: "100vh",
       }}
     >
-      <CustomAppbar position = 'static'/>
+      <CustomAppbar position="static" />
       <LoginForm />
     </Box>
   );
