+++ client/src/pages/register-page.js
@@ -1,7 +1,6 @@
 import React from "react";
 import RegisterForm from "../components/UI/Forms/register-form";
 import { Box } from "@mui/material";
-import Header from "../components/UI/Navbar/Header";
 import CustomAppbar from "../components/UI/Appbar/custom-appbar";
 
 export const RegisterPages = () => {
